Added status updation method

+++ controllers/feed.js
@@ -3,7 +3,8 @@ const path = require('path');
 
 const {validationResult} = require('express-validator');
 
-const Post = require('../models/post')
+const Post = require('../models/post');
+const User = require('../models/user')
 
 exports.getPosts = (req, res, next)=> {
     const currentPage = req.query.page || 1;
@@ -21,13 +22,11 @@ exports.getPosts = (req, res, next)=> {
             res.status(200).json({message: 'Fetched Posts Successfully.', posts: posts, totalItems: totalItems})
         })
         .catch(err => {
-            if(err.statusCode){
+            if(!err.statusCode){
                 err.statusCode = 500;
             }
             next(err);
         });
-
-    
 };
 
 exports.createPost = (req, res, next) => {
@@ -45,21 +44,31 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path.replace("\\","/");
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl:imageUrl,
-        creator: {name: 'Vishal'},
+        creator: req.userId
     });
     post.save()
     .then(result => {
+        return User.findById(req.userId);
+    })
+    .then(user => {
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+    })
+    .then(result => {
         res.status(201).json({
             message: 'Post Created Successfully!',
-            post: result
+            post: post,
+            creator: { _id: creator._id, name: creator.name }
         });
     })
     .catch(err => {
-        if(err.statusCode){
+        if(!err.statusCode){
             err.statusCode = 500;
         }
         next(err);
@@ -79,7 +88,7 @@ exports.getPost = (req, res, next) => {
             res.status(200).json({message: 'Post fetched.', post:post})
         })
         .catch(err => {
-            if(err.statusCode){
+            if(!err.statusCode){
                 err.statusCode = 500;
             }
             next(err);
@@ -112,6 +121,11 @@ exports.updatePost =(req, res, next) => {
                 error.statusCode=404;
                 throw error;
             }
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!!');
+                error.statusCode=403;
+                throw error;
+            }
             if(imageUrl !==post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -124,7 +138,7 @@ exports.updatePost =(req, res, next) => {
             res.status(200).json({message: 'Post updated!', post: result});
         })
         .catch(err => {
-            if(err.statusCode){
+            if(!err.statusCode){
                 err.statusCode = 500;
             }
             next(err);
@@ -140,16 +154,27 @@ exports.deletePost= (req, res, next) =>{
                 error.statusCode=404;
                 throw error;
             }
-            //check logged in user
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!!');
+                error.statusCode=403;
+                throw error;
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndDelete(postId);
         })
         .then(result => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+            
+        })
+        .then(result => {
             res.status(200).json({message: 'Deleted Post.'})
         })
         .catch(err => {
-            if(err.statusCode){
+            if(!err.statusCode){
                 err.statusCode = 500;
             }
             next(err);
